requirements.txt
Django==5.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
boto3==1.34.22
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
python-docx==1.1.0
celery==5.3.6
redis==5.0.1
